refactor(cvtservice): extract consumer_registry.go from god-file
Extract RegisterConsumer + ListConsumers + DeregisterConsumer into a
dedicated file. Pure file movement, no behavior change.

Part 2 of 4 in the validator_service.go split (issue #107).
